Refactor API startup composition
